1. Change query into Eloquent 2. Remove php code from blade file 3. Send email on successfully registration using queue and job

+++ resources/views/users/index.blade.php
@@ -1,15 +1,12 @@
 @extends('layouts.app')
 @section('content')
-<?php
-    $array = Auth::user()->getRoleNames();
-    $role = trim($array,'["]');
-?>
+
 <div class="row">
     <div class="col-lg-12 margin-tb">
         <div class="pull-left">
             <h2>Users Management</h2>
         </div>
-        @if($role == "Admin")
+        @if($user_role == "Admin")
         <div class="pull-right">
             <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
         </div>
@@ -32,7 +29,7 @@
         <th width="280px">Action</th>
     </tr>
     
-    @if($role == "Admin")
+    @if($user_role == "Admin")
         @foreach ($data as $key => $user)
         <tr>
             <td>{{ ++$i }}</td>
@@ -48,9 +45,11 @@
             <td>
             <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
             <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+            @if($user_role == "Admin")
                 {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                 {!! Form::close() !!}
+                @endif
             </td>
         </tr>
         @endforeach
